+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add a Travis check to make sure version update
 - Add metrics for Selenium initialization metrics for integration tests
+- [Demo] Show video WebRTC stats and attendeeId on video tile hover
 
 ### Changed
 - Update test results to Sauce Labs before emitting CloudWatch metrics for integration tests